functions/move_file/requirements.txt
google-cloud-storage==1.13.1
google-cloud-firestore==0.30.1